슬랙 메시지 전송의 제어권을 역전시키도록 변경 (IoC)
